third_party/blink/web_tests/external/wpt/html/browsers/browsing-the-web/navigating-across-documents/010.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Link with onclick form submit to javascript url with delayed document.write and href navigation 
  assert_equals: expected true but got false
Harness: the test ran to completion.
